+++ src/num/geom/shape/point/methods.rs
@@ -2,48 +2,35 @@
 //
 //!
 //
+// TOC
+// - Point
+// - Point2d
+// - Point3d
+// - impl_point!
+// - Points
 
 #[cfg(feature = "linear")]
 use crate::Vector;
-use crate::{Point, Point2d, Point3d};
+use crate::{Point, Point2d, Point3d, Points};
 
 #[rustfmt::skip]
 impl<T, const D: usize> Point<T, D> {
-    /// Returns a new `Point` from the given `coords` array.
-    #[must_use]
-    pub const fn new(coords: [T; D]) -> Self {
-        Self { coords }
-    }
+    #[must_use] /// Returns a new `Point` from the given `coords` array.
+    pub const fn new(coords: [T; D]) -> Self { Self { coords } }
 
-    /// Consumes this `Point` and converts it into a `Vector`.
-    #[must_use]
-    #[cfg(feature = "linear")]
-    #[cfg_attr(nightly_doc, doc(cfg(feature = "linear")))]
-    pub fn into_vector(self) -> Vector<T, D> {
-        Vector::new(self.coords)
-    }
-    /// Converts this `Point` to a `Vector`, keeping `self` intact.
-    #[must_use]
-    #[cfg(feature = "linear")]
-    #[cfg_attr(nightly_doc, doc(cfg(feature = "linear")))]
-    pub const fn to_vector(self) -> Vector<T, D> where T: Copy {
-        Vector::new(self.coords)
-    }
+    #[must_use] /// Consumes this `Point` and converts it into a `Vector`.
+    #[cfg(feature = "linear")] #[cfg_attr(nightly_doc, doc(cfg(feature = "linear")))]
+    pub fn into_vector(self) -> Vector<T, D> { Vector::new(self.coords) }
+    #[must_use] /// Converts this `Point` to a `Vector`, keeping `self` intact.
+    #[cfg(feature = "linear")] #[cfg_attr(nightly_doc, doc(cfg(feature = "linear")))]
+    pub const fn to_vector(self) -> Vector<T, D> where T: Copy { Vector::new(self.coords) }
 
-    /// Creates a `Point` from a `Vector`.
-    #[must_use]
-    #[cfg(feature = "linear")]
-    #[cfg_attr(nightly_doc, doc(cfg(feature = "linear")))]
-    pub fn from_vector(v: Vector<T, D>) -> Self {
-        Self::new(v.coords)
-    }
-    /// Creates a `Point` from a constant `Vector`.
-    #[must_use]
-    #[cfg(feature = "linear")]
-    #[cfg_attr(nightly_doc, doc(cfg(feature = "linear")))]
-    pub const fn from_vector_const(v: Vector<T, D>) -> Self where T: Copy {
-        Self::new(v.coords)
-    }
+    #[must_use] /// Creates a `Point` from a `Vector`.
+    #[cfg(feature = "linear")] #[cfg_attr(nightly_doc, doc(cfg(feature = "linear")))]
+    pub fn from_vector(v: Vector<T, D>) -> Self { Self::new(v.coords) }
+    #[must_use] /// Creates a `Point` from a constant `Vector`.
+    #[cfg(feature = "linear")] #[cfg_attr(nightly_doc, doc(cfg(feature = "linear")))]
+    pub const fn from_vector_const(v: Vector<T, D>) -> Self where T: Copy { Self::new(v.coords) }
 }
 
 /* manual impls for specific dimensionalities */
@@ -51,58 +38,43 @@ impl<T, const D: usize> Point<T, D> {
 #[rustfmt::skip]
 impl<T> Point2d<T> {
     /// Returns a copy of the first dimension `x`.
-    #[must_use]
-    pub const fn x(self) -> T where T: Copy { self.coords[0] }
+    #[must_use] pub const fn x(self) -> T where T: Copy { self.coords[0] }
     /// Returns a copy of the second dimension `y`.
-    #[must_use]
-    pub const fn y(self) -> T where T: Copy { self.coords[1] }
+    #[must_use] pub const fn y(self) -> T where T: Copy { self.coords[1] }
 
     /// Returns a shared reference to the first dimension `x`.
-    #[must_use]
-    pub const fn x_ref(&self) -> &T { &self.coords[0] }
+    #[must_use] pub const fn x_ref(&self) -> &T { &self.coords[0] }
     /// Returns a shared reference to the second dimension `y`.
-    #[must_use]
-    pub const fn y_ref(&self) -> &T { &self.coords[1] }
+    #[must_use] pub const fn y_ref(&self) -> &T { &self.coords[1] }
 
     /// Returns an exclusive reference to the first dimension `x`.
-    #[must_use]
-    pub fn x_mut(&mut self) -> &mut T { &mut self.coords[0] }
+    #[must_use] pub const fn x_mut(&mut self) -> &mut T { &mut self.coords[0] }
     /// Returns an exclusive reference to the second dimension `y`.
-    #[must_use]
-    pub fn y_mut(&mut self) -> &mut T { &mut self.coords[1] }
+    #[must_use] pub const fn y_mut(&mut self) -> &mut T { &mut self.coords[1] }
 }
 
 #[rustfmt::skip]
 impl<T> Point3d<T> {
     /// Returns a copy of the first dimension `x`.
-    #[must_use]
-    pub const fn x(self) -> T where T: Copy { self.coords[0] }
+    #[must_use] pub const fn x(self) -> T where T: Copy { self.coords[0] }
     /// Returns a copy of the second dimension `y`.
-    #[must_use]
-    pub const fn y(self) -> T where T: Copy { self.coords[1] }
+    #[must_use] pub const fn y(self) -> T where T: Copy { self.coords[1] }
     /// Returns a copy of the third dimension `z`.
-    #[must_use]
-    pub const fn z(self) -> T where T: Copy { self.coords[2] }
+    #[must_use] pub const fn z(self) -> T where T: Copy { self.coords[2] }
 
     /// Returns a shared reference to the first dimension `x`.
-    #[must_use]
-    pub const fn x_ref(&self) -> &T { &self.coords[0] }
+    #[must_use] pub const fn x_ref(&self) -> &T { &self.coords[0] }
     /// Returns a shared reference to the second dimension `y`.
-    #[must_use]
-    pub const fn y_ref(&self) -> &T { &self.coords[1] }
+    #[must_use] pub const fn y_ref(&self) -> &T { &self.coords[1] }
     /// Returns a shared reference to the third dimension `z`.
-    #[must_use]
-    pub const fn z_ref(&self) -> &T { &self.coords[2] }
+    #[must_use] pub const fn z_ref(&self) -> &T { &self.coords[2] }
 
     /// Returns an exclusive reference to the first dimension `x`.
-    #[must_use]
-    pub fn x_mut(&mut self) -> &mut T { &mut self.coords[0] }
+    #[must_use] pub const fn x_mut(&mut self) -> &mut T { &mut self.coords[0] }
     /// Returns an exclusive reference to the second dimension `y`.
-    #[must_use]
-    pub fn y_mut(&mut self) -> &mut T { &mut self.coords[1] }
+    #[must_use] pub const fn y_mut(&mut self) -> &mut T { &mut self.coords[1] }
     /// Returns an exclusive reference to the third dimension `z`.
-    #[must_use]
-    pub fn z_mut(&mut self) -> &mut T { &mut self.coords[2] }
+    #[must_use] pub const fn z_mut(&mut self) -> &mut T { &mut self.coords[2] }
 }
 
 /// helper for implementing methods on `Point`.
@@ -111,9 +83,11 @@ impl<T> Point3d<T> {
 /// $cap:  the capability feature that enables the given implementation. E.g "_int_i8".
 macro_rules! impl_point {
     () => {
-        impl_point![sint i8:"_int_i8", i16:"_int_i16", i32:"_int_i32",
+        impl_point![sint
+            i8:"_int_i8", i16:"_int_i16", i32:"_int_i32",
             i64:"_int_i64", i128:"_int_i128", isize:"_int_isize"];
-        impl_point![uint u8:"_int_u8", u16:"_int_u16", u32:"_int_u32",
+        impl_point![uint
+            u8:"_int_u8", u16:"_int_u16", u32:"_int_u32",
             u64:"_int_u64", u128:"_int_u128", usize:"_int_usize"];
         impl_point![float f32:"_float_f32", f64:"_float_f64"];
     };
@@ -180,4 +154,12 @@ macro_rules! impl_point {
         }
     };
 }
+use impl_point;
 impl_point!();
+
+#[rustfmt::skip]
+impl<T, const D: usize, const N: usize> Points<T, D, N> {
+    #[must_use]
+    /// Returns new `Points` from the given `coords` array.
+    pub const fn new(points: [Point<T, D>; N]) -> Self { Self { array: points } }
+}